Remove COPY ifdef from crc32 (v)pclmulqdq.
